chore: update start script for new cli path
