+++ tsconfig.json
@@ -13,5 +13,5 @@
     "resolveJsonModule": true,
     "typeRoots": ["./types", "./node_modules/@types"],
   },
-  "include": ["src/**/*.ts", "types/**/*.d.ts"]
+  "include": ["src/**/*.ts", "bin/**/*.ts", "types/**/*.d.ts"]
 }
